+++ changelog.md
@@ -1,5 +1,14 @@
 # Changelog
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+### Added
+- Extended interval aminoglycoside calculator [#79](https://github.com/pharmot/multipurpose-calculator/issues/79)
+- addHoursToDate function in Utility Module
+- Vancomycin peak timing calculator [#66](https://github.com/pharmot/multipurpose-calculator/issues/66)
+
+### Changed
+- Updated tints and shades of VMFH colors [#77](https://github.com/pharmot/multipurpose-calculator/issues/77)
+- Split left column (patient info) into its own partial html file
 
 ## [1.1.0] - 2023-01-30